WIP: move CommitChanges out of ducklake transaction
