+++ src/storage/ducklake_transaction.cpp
@@ -1125,22 +1125,6 @@ DuckLakeTableInfo DuckLakeTransaction::GetNewTable(DuckLakeCommitState &commit_s
 	return table_entry;
 }
 
-struct NewTableInfo {
-	vector<DuckLakeTableInfo> new_tables;
-	vector<DuckLakeViewInfo> new_views;
-	vector<DuckLakePartitionInfo> new_partition_keys;
-	vector<DuckLakeTagInfo> new_tags;
-	vector<DuckLakeColumnTagInfo> new_column_tags;
-	vector<DuckLakeDroppedColumn> dropped_columns;
-	vector<DuckLakeNewColumn> new_columns;
-	vector<DuckLakeTableInfo> new_inlined_data_tables;
-	vector<DuckLakeSortInfo> new_sort_keys;
-};
-
-struct NewMacroInfo {
-	vector<DuckLakeMacroInfo> new_macros;
-};
-
 void CancelOrDropField(TableIndex table_id, FieldIndex field_id, NewTableInfo &result,
                        unordered_map<idx_t, idx_t> &txn_added_fields) {
 	auto it = txn_added_fields.find(field_id.index);
@@ -1692,11 +1676,6 @@ DuckLakeFileInfo DuckLakeTransaction::GetNewDataFile(const DuckLakeDataFile &fil
 	return data_file;
 }
 
-struct NewDataInfo {
-	vector<DuckLakeFileInfo> new_files;
-	vector<DuckLakeInlinedDataInfo> new_inlined_data;
-};
-
 NewDataInfo DuckLakeTransaction::GetNewDataFiles(string &batch_query, DuckLakeCommitState &commit_state,
                                                  optional_ptr<vector<DuckLakeGlobalStatsInfo>> stats) {
 	NewDataInfo result;
@@ -1852,10 +1831,6 @@ DuckLakeTransaction::GetNewDeleteFiles(const DuckLakeCommitState &commit_state,
 	return result;
 }
 
-struct NewNameMapInfo {
-	vector<DuckLakeColumnMappingInfo> new_column_mappings;
-};
-
 void ConvertNameMapColumn(const DuckLakeNameMapEntry &name_map_entry, MappingIndex map_id, idx_t &column_idx,
                           DuckLakeColumnMappingInfo &result, optional_idx parent_idx = optional_idx()) {
 	auto column_id = column_idx++;
@@ -1920,205 +1895,6 @@ DuckLakeTransaction::GetNewInlinedDeletes(DuckLakeCommitState &commit_state) con
 		}
 	}
 	return result;
-}
-
-struct CompactionInformation {
-	vector<DuckLakeCompactedFileInfo> compacted_files;
-	vector<DuckLakeFileInfo> new_files;
-};
-
-string DuckLakeTransaction::CommitChanges(DuckLakeCommitState &commit_state,
-                                          TransactionChangeInformation &transaction_changes,
-                                          optional_ptr<vector<DuckLakeGlobalStatsInfo>> stats) {
-	auto &commit_snapshot = commit_state.commit_snapshot;
-	auto &dropped_tables = state->dropped_tables;
-	auto &renamed_tables = state->renamed_tables;
-	auto &dropped_views = state->dropped_views;
-	auto &renamed_views = state->renamed_views;
-	auto &dropped_scalar_macros = state->dropped_scalar_macros;
-	auto &dropped_table_macros = state->dropped_table_macros;
-	auto &dropped_schemas = state->dropped_schemas;
-	auto &new_schemas = state->new_schemas;
-	auto &new_tables = state->new_tables;
-	auto &new_scalar_macros = state->new_scalar_macros;
-	auto &new_table_macros = state->new_table_macros;
-	auto &local_changes = state->local_changes;
-	auto &flushed_inlined_tables = state->flushed_inlined_tables;
-	auto &dropped_files = state->dropped_files;
-
-	ducklake_catalog.EnsureCommitInfoProvided(commit_info);
-
-	string batch_queries;
-	// drop entries
-	if (!dropped_tables.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropTables(dropped_tables, false);
-	}
-
-	if (!renamed_tables.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropTables(renamed_tables, true);
-	}
-
-	if (!dropped_views.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropViews(dropped_views, false);
-	}
-
-	if (!renamed_views.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropViews(renamed_views, true);
-	}
-
-	if (!dropped_scalar_macros.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropMacros(dropped_scalar_macros);
-	}
-
-	if (!dropped_table_macros.empty()) {
-		batch_queries += DuckLakeMetadataManager::DropMacros(dropped_table_macros);
-	}
-	if (!dropped_schemas.empty()) {
-		set<SchemaIndex> dropped_schema_ids;
-		for (auto &entry : dropped_schemas) {
-			dropped_schema_ids.insert(entry.first);
-		}
-		batch_queries += DuckLakeMetadataManager::DropSchemas(dropped_schema_ids);
-	}
-	// write new schemas
-	vector<DuckLakeSchemaInfo> new_schemas_result;
-	if (new_schemas) {
-		new_schemas_result = GetNewSchemas(commit_state);
-		vector<DuckLakePath> resolved_schema_paths;
-		resolved_schema_paths.reserve(new_schemas_result.size());
-		for (auto &schema : new_schemas_result) {
-			resolved_schema_paths.push_back(metadata_manager->GetRelativePath(schema.path));
-		}
-		batch_queries += DuckLakeMetadataManager::WriteNewSchemas(new_schemas_result, resolved_schema_paths);
-	}
-
-	// write new tables
-	vector<DuckLakeTableInfo> new_tables_result;
-	vector<DuckLakeTableInfo> new_inlined_data_tables_result;
-	if (!new_tables.empty()) {
-		auto result = GetNewTables(commit_state, transaction_changes);
-		vector<DuckLakePath> resolved_table_paths;
-		resolved_table_paths.reserve(result.new_tables.size());
-		for (auto &table : result.new_tables) {
-			resolved_table_paths.push_back(
-			    metadata_manager->GetRelativePath(table.schema_id, table.path, new_schemas_result));
-		}
-		batch_queries += DuckLakeMetadataManager::WriteNewTables(result.new_tables, resolved_table_paths);
-		auto existing_catalog = metadata_manager->GetCatalogForSnapshot(commit_snapshot);
-		batch_queries +=
-		    DuckLakeMetadataManager::WriteNewPartitionKeys(existing_catalog.partitions, result.new_partition_keys);
-		batch_queries += DuckLakeMetadataManager::WriteNewViews(result.new_views);
-		batch_queries += DuckLakeMetadataManager::WriteNewTags(result.new_tags);
-		batch_queries += DuckLakeMetadataManager::WriteNewColumnTags(result.new_column_tags);
-		batch_queries += DuckLakeMetadataManager::WriteDroppedColumns(result.dropped_columns);
-		batch_queries += DuckLakeMetadataManager::WriteNewColumns(result.new_columns);
-		batch_queries += metadata_manager->WriteNewInlinedTables(commit_snapshot, result.new_inlined_data_tables);
-		batch_queries += DuckLakeMetadataManager::WriteNewSortKeys(existing_catalog.sorts, result.new_sort_keys);
-		new_tables_result = result.new_tables;
-		new_inlined_data_tables_result = result.new_inlined_data_tables;
-	}
-
-	if (!new_scalar_macros.empty() || !new_table_macros.empty()) {
-		auto result = GetNewMacros(commit_state, transaction_changes);
-		batch_queries += DuckLakeMetadataManager::WriteNewMacros(result.new_macros);
-	}
-
-	// write new name maps
-	if (!new_name_maps.name_maps.empty()) {
-		auto result = GetNewNameMaps(commit_state);
-		batch_queries += DuckLakeMetadataManager::WriteNewColumnMappings(result.new_column_mappings);
-	}
-
-	// write new data / data files
-	bool has_table_data_changes = local_changes.HasChanges();
-	if (has_table_data_changes) {
-		auto result = GetNewDataFiles(batch_queries, commit_state, stats);
-		batch_queries += metadata_manager->WriteNewDataFiles(commit_snapshot, result.new_files, new_tables_result,
-		                                                     new_schemas_result);
-		batch_queries += metadata_manager->WriteNewInlinedData(commit_snapshot, result.new_inlined_data,
-		                                                       new_tables_result, new_inlined_data_tables_result);
-	}
-
-	// in case of a retry, we generate the deletion of inlined data from the tables
-	if (!flushed_inlined_tables.empty()) {
-		batch_queries += DuckLakeMetadataManager::GenerateDeleteFlushedInlinedData(flushed_inlined_tables);
-	}
-
-	// drop data files
-	if (!dropped_files.empty()) {
-		set<DataFileIndex> dropped_indexes;
-		for (auto &entry : dropped_files) {
-			dropped_indexes.insert(entry.second);
-		}
-		batch_queries += DuckLakeMetadataManager::DropDataFiles(dropped_indexes);
-	}
-
-	if (has_table_data_changes) {
-		// write new delete files
-		vector<DuckLakeOverwrittenDeleteFile> overwritten_delete_files;
-		auto file_list = GetNewDeleteFiles(commit_state, overwritten_delete_files);
-		vector<DuckLakePath> resolved_overwritten_paths;
-		resolved_overwritten_paths.reserve(overwritten_delete_files.size());
-		for (auto &file : overwritten_delete_files) {
-			resolved_overwritten_paths.push_back(metadata_manager->GetRelativePath(file.path));
-		}
-		batch_queries +=
-		    DuckLakeMetadataManager::DeleteOverwrittenDeleteFiles(overwritten_delete_files, resolved_overwritten_paths);
-		vector<DuckLakePath> resolved_delete_paths;
-		resolved_delete_paths.reserve(file_list.size());
-		for (auto &file : file_list) {
-			resolved_delete_paths.push_back(
-			    metadata_manager->GetRelativePath(file.table_id, file.path, new_tables_result, new_schemas_result));
-		}
-		batch_queries += DuckLakeMetadataManager::WriteNewDeleteFiles(file_list, resolved_delete_paths);
-
-		// write new inlined deletes (for inlined data tables)
-		auto inlined_deletes = GetNewInlinedDeletes(commit_state);
-		batch_queries += DuckLakeMetadataManager::WriteNewInlinedDeletes(inlined_deletes);
-
-		// write new inlined file deletes (for parquet files)
-		auto inlined_file_deletes = GetNewInlinedFileDeletes(commit_state);
-		batch_queries += metadata_manager->WriteNewInlinedFileDeletes(commit_snapshot, inlined_file_deletes);
-
-		// write compactions
-		auto compaction_merge_adjacent_changes =
-		    GetCompactionChanges(commit_state, CompactionType::MERGE_ADJACENT_TABLES);
-		vector<DuckLakePath> resolved_merge_paths;
-		resolved_merge_paths.reserve(compaction_merge_adjacent_changes.compacted_files.size());
-		for (auto &compaction : compaction_merge_adjacent_changes.compacted_files) {
-			resolved_merge_paths.push_back(metadata_manager->GetRelativePath(compaction.path));
-		}
-		batch_queries +=
-		    DuckLakeMetadataManager::WriteCompactions(compaction_merge_adjacent_changes.compacted_files,
-		                                              CompactionType::MERGE_ADJACENT_TABLES, resolved_merge_paths);
-		batch_queries += metadata_manager->WriteNewDataFiles(
-		    commit_snapshot, compaction_merge_adjacent_changes.new_files, new_tables_result, new_schemas_result);
-
-		auto compaction_rewrite_delete_changes = GetCompactionChanges(commit_state, CompactionType::REWRITE_DELETES);
-		batch_queries += metadata_manager->WriteNewDataFiles(
-		    commit_snapshot, compaction_rewrite_delete_changes.new_files, new_tables_result, new_schemas_result);
-		// REWRITE_DELETES ignores resolved_paths; pass empty.
-		batch_queries += DuckLakeMetadataManager::WriteCompactions(
-		    compaction_rewrite_delete_changes.compacted_files, CompactionType::REWRITE_DELETES, vector<DuckLakePath>());
-	}
-
-	// Tracking for tables that had schema changes
-	set<TableIndex> tables_with_schema_changes;
-	for (auto &table_id : transaction_changes.altered_tables) {
-		if (!table_id.IsTransactionLocal() &&
-		    transaction_changes.altered_tables_with_schema_version_changes.find(table_id) !=
-		        transaction_changes.altered_tables_with_schema_version_changes.end()) {
-			tables_with_schema_changes.insert(table_id);
-		}
-	}
-	for (auto &new_table : new_tables_result) {
-		if (!new_table.id.IsTransactionLocal()) {
-			tables_with_schema_changes.insert(new_table.id);
-		}
-	}
-	batch_queries += DuckLakeMetadataManager::InsertNewSchema(commit_snapshot, tables_with_schema_changes);
-
-	return batch_queries;
 }
 
 CompactionInformation DuckLakeTransaction::GetCompactionChanges(DuckLakeCommitState &commit_state,
